AddedDataValidation

+++ Controllers/FlightController.cs
@@ -16,26 +16,28 @@ namespace SystemZarzadzaniaLotami.Controllers
     [ApiController]
     public class FlightController : ControllerBase
     {
-        private readonly DatabaseContext db;
         private readonly IFlightRepository flightRepository;
 
         public FlightController(DatabaseContext context, IFlightRepository flightRepo)
         {
             flightRepository = flightRepo;
-            db = context;
         }
 
         [HttpGet]
         public async Task<ActionResult<IEnumerable<Flight>>> GetFlights()
         {
+            if(!ModelState.IsValid) { return BadRequest(ModelState); }
+            
             var flights = await flightRepository.GetFlightsAsync();
             var flightsDTO = flights.Select(s=>s.ToFlightsDTO());
             return Ok(flights);
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<ActionResult<Flight>> GetFlight(int id)
         {
+            if (!ModelState.IsValid) { return BadRequest(ModelState); }
+
             var flight = await flightRepository.GetFlightAsync(id);
 
             if (flight == null) 
@@ -49,16 +51,22 @@ namespace SystemZarzadzaniaLotami.Controllers
         [HttpPost]
         public async Task<ActionResult<Flight>> AddFlight([FromBody] AddFlightDTO flightDTO)
         {
+            if (!ModelState.IsValid) { return BadRequest(ModelState); }
+
             var flightModel = flightDTO.ToFlightFromAddDTO();
+            Console.WriteLine("contr iddddddddddd= " + flightModel.Id);
+
             await flightRepository.AddFlightAsync(flightModel);
 
             return CreatedAtAction(nameof(GetFlight), new { id = flightModel.Id }, flightModel.ToFlightsDTO());
         }
 
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> UpdateFlight([FromRoute] int id, [FromBody] UpdateFlightDTO updateDTO)
         {
+            if (!ModelState.IsValid) { return BadRequest(ModelState); }
+
             var flightModel = await flightRepository.UpdateFlightAsync(id,updateDTO);
             if (flightModel == null)
             {
@@ -69,9 +77,11 @@ namespace SystemZarzadzaniaLotami.Controllers
         }
 
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> DeleteFlight([FromRoute] int id)
         {
+            if (!ModelState.IsValid) { return BadRequest(ModelState); }
+
             var flightModel = await flightRepository.DeleteFlightAsync(id);
             if (flightModel == null)
             {
